Refactor domain
