+++ websocket/app/Cargo.toml
@@ -25,6 +25,5 @@ dotenv.workspace = true
 envy.workspace = true
 flow-websocket-services.workspace = true
 flow-websocket-infra.workspace = true
-flow-websocket-domain.workspace = true
 bb8.workspace = true
 bb8-redis.workspace = true
