--- websocket/crates/domain/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-description = "Re:Earth Flow Websocket Domain library"
-name = "flow-websocket-domain"
-
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-
-async-trait.workspace = true
-chrono.workspace = true
-mockall.workspace = true
-redis.workspace = true
-serde.workspace = true
-thiserror.workspace = true
-tokio.workspace = true
-tracing.workspace = true
-uuid.workspace = true
