--- websocket/crates/domain/src/lib.rs
@@ -1,9 +1,0 @@
-pub mod editing_session;
-pub mod repository;
-pub mod types;
-pub use types::project;
-pub use types::snapshot;
-pub use types::user;
-mod utils;
-pub use editing_session::ProjectEditingSession;
-pub use utils::calculate_diff;
