+++ websocket/crates/infra/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod persistence;
+pub mod types;
+pub mod utils;
