+++ websocket/crates/infra/src/persistence/project_repository.rs
@@ -1,21 +1,18 @@
+use crate::generate_id;
 use crate::persistence::gcs::gcs_client::{GcsClient, GcsError};
-use async_trait::async_trait;
-use flow_websocket_domain::generate_id;
-use flow_websocket_domain::project::Project;
-
 use crate::persistence::local_storage::LocalClient;
-use flow_websocket_domain::editing_session::ProjectEditingSession;
-use flow_websocket_domain::repository::{
-    ProjectEditingSessionImpl, ProjectImpl, ProjectSnapshotImpl,
-};
-use flow_websocket_domain::snapshot::ProjectSnapshot;
+use crate::types::project::Project;
+use crate::types::snapshot::ProjectSnapshot;
+use async_trait::async_trait;
 use serde_json;
 use std::io;
 use std::path::PathBuf;
 use std::sync::Arc;
 use thiserror::Error;
 
+use super::editing_session::ProjectEditingSession;
 use super::local_storage::LocalStorageError;
+use super::repository::{ProjectEditingSessionImpl, ProjectImpl, ProjectSnapshotImpl};
 use super::StorageClient;
 use bb8::Pool;
 use bb8_redis::RedisConnectionManager;
