+++ websocket/crates/infra/src/types/snapshot.rs
