+++ websocket/crates/infra/src/utils.rs
@@ -1,0 +1,6 @@
+#[macro_export]
+macro_rules! generate_id {
+    ($prefix:expr) => {
+        format!("{}{}", $prefix, uuid::Uuid::new_v4())
+    };
+}
